Fix multi-query resultset iteration
This uses an ugly hack to keep the standard rows iterator from closing the
HTTP body when it reaches the end: It sets the body to nil.

I think there's room now to consolidate the `rows` and `multiQueriesRows`.
My initial thought was to expand `rows`, but now that I have two
implementations, I believe it would make more sense instead to consolidate
everything in to `multiQueriesRows`.  It already has a few special cases
anyway.