Drop cookie if we get a 401 after authenticating using one.
That way we shouldn't end up with a situation where the cookie is set
but invalid for too long.

In theory a user can retry on 401.  We avoid doing this as part of the
RoundTripper interface because that would require storing the body.  And
we'd need to do that in anticipation of a failure, which is too much of
a burden just in case we might fail.