Test that other cookies set by the server are not wiped out.
Note that this exposes an interesting property of the auth mechanism.
On the first request after auth, if extra cookies were sent, they won't
be sent on that request.  They will only be applied to following
requests.  This is due to the fact that the request has been 'baked'
before the auth request/response and we only copy over the auth cookie
onto that request.

This test captures that (and has a comment to note that behaviour).
That's existing behaviour, and at this point I figure it's just worth
noting.